Thanks everybody for the hard work and testing on this.  HUZZAH!

* add client assets to .gitignore

* Initial attempt at 64-bit hacking...

* Finished with ODT_CreateObject() complete.

* OD2_CreateObject passed.

* First Running Example!
Still crashes in some spots.

* Tribad/64bit (#73)

* Tribad's changes to the 64-Bit port.

* Tribad's initial CMake files to get my own environment up and running.

* Introduced  struct to come around the nasty dword array. for animation stuff

* Use the systems glew.h on linux.

* Various formatting and such.

* Removing CMake since it has been branched for further discussion

* Fixing indentation and line endings on obj_odt.cpp

* Fixing indentation and line endings in od_odt.h

* Revert linux to included glew

* copy BoundingSphere from 32bit obj.

* Update sl_msg.cpp

Fixed the headless bot output, it was running vsprintf twice before and that causes a bigger problem on 64bit.

* Xcode project changes for 64 bit

* Fixes for arm64: clean up writestring function pointer code.

* Fix compilation when PARSEC_DEBUG is defined

* visual studio: add x64 targets

* Fix compilation on windows/x64

* visual studio: update gitignore for x64

* windows: fix SYSs_AcquireScriptPath on 64 bit systems

* 64 bit: fix assertion in OD2_CreateObject

* visual studio: debug builds enable assertions via PARSEC_DEBUG

* Remove obsolete comment

+++ src/libparsec/sw_path.cpp
@@ -133,9 +133,9 @@ int SYSs_AcquireScriptPath( char *path, int comtype, char *prefix )
 
 	// open script directory
 	struct _finddata_t  c_file;
-	long hFile = _findfirst( path, &c_file );
+	intptr_t hFile = _findfirst( path, &c_file );
 
-	if ( hFile != -1L ) {
+	if ( hFile != -1LL ) {
 
 		int countbase = num_external_commands;
